requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.1.0
billiard==4.2.0
black==24.8.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.9
django-celery-results==2.5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.22.1
flower==2.0.1
humanize==4.10.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pip-check-reqs==2.5.3
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.19.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
steptools==20.5
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
vulture==2.11
wcwidth==0.2.13
psycopg2==2.9.9
psycopg2-binary==2.9.9
